refactor(controllers): lift shared reconcile config into base.Options
The KS, HR, and RS controllers each duplicated the same five
reconcile-shaping fields (Filter, ParentOf, RenderTracker, Existence,
PreflightFailure) on their option structs and the same five-setter
Configure preamble. Hoist both into base.Options + base.Controller.Configure;
each controller now embeds base.Options and forwards to it before assigning
its own kind-specific fields (SelfProduces, AllowMissingSecrets/Producers,
RawSink).

Pure refactor: `flate build all` is byte-identical on k8s-gitops
(2,382,360 bytes). The source controller stays separate — it is
fetch-only and shares just Filter.
